Simplifying PerdioVar  Plan split it into:   LazyObjects   DistributedVariable   DistributedFuture (maybe shared with DistributedVariable)
Step1:
  use virtual functions to access PerdioVars
  OldPerdioVar is the beast from before, but uses virtual functions now

TODO:
  split up OldPerdioVar

use %ld instead of (int) cast to fix compiler warnings

+++ platform/emulator/print.cc
@@ -392,7 +392,7 @@ void GenCVariable::printStream(ostream &stream, int depth)
       break;
 
   default:
-    stream << " unknown type: " << getType();
+    stream << " unknown type: " << (int) getType();
     break;
   }
   stream << ">";
@@ -471,7 +471,7 @@ void GenCVariable::printLongStream(ostream &stream, int depth, int offset)
     stream << indent(offset) << "<PerdioVariable *" << this << ">" << endl;
       break;
   default:
-    stream << indent(offset) << " unknown type: " << getType() << endl;
+    stream << indent(offset) << " unknown type: " << (int) getType() << endl;
     break;
   }
